Make ephemeral dialog a side effect
