+++ android/space/src/main/kotlin/net/twisterrob/inventory/android/space/ManageSpaceUiEffect.kt
@@ -13,4 +13,6 @@ internal sealed class ManageSpaceUiEffect {
 	data class PickDumpAllDataTarget(
 		val fileName: String,
 	) : ManageSpaceUiEffect()
+
+	object PickVacuumDatabaseIncrementalBytes : ManageSpaceUiEffect()
 }
